SettingSceneをChooseMusicSceneから分離 settingをシングルトンに
