+++ 音ゲー（仮）.xcodeproj/xcshareddata/xcschemes/音ゲー（仮）.xcscheme
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
    LastUpgradeVersion = "1010"
-   version = "1.3">
+   version = "1.7">
    <BuildAction
       parallelizeBuildables = "YES"
       buildImplicitDependencies = "YES">
@@ -28,36 +28,7 @@
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       shouldUseLaunchSchemeArgsEnv = "YES">
       <Testables>
-         <TestableReference
-            skipped = "NO">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "0FFD1B591F5E6D2B0007E555"
-               BuildableName = "&#38899;&#12465;&#12441;&#12540;&#65288;&#20206;&#65289;Tests.xctest"
-               BlueprintName = "&#38899;&#12465;&#12441;&#12540;&#65288;&#20206;&#65289;Tests"
-               ReferencedContainer = "container:&#38899;&#12465;&#12441;&#12540;&#65288;&#20206;&#65289;.xcodeproj">
-            </BuildableReference>
-         </TestableReference>
-         <TestableReference
-            skipped = "NO">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "0FFD1B641F5E6D2B0007E555"
-               BuildableName = "&#38899;&#12465;&#12441;&#12540;&#65288;&#20206;&#65289;UITests.xctest"
-               BlueprintName = "&#38899;&#12465;&#12441;&#12540;&#65288;&#20206;&#65289;UITests"
-               ReferencedContainer = "container:&#38899;&#12465;&#12441;&#12540;&#65288;&#20206;&#65289;.xcodeproj">
-            </BuildableReference>
-         </TestableReference>
       </Testables>
-      <MacroExpansion>
-         <BuildableReference
-            BuildableIdentifier = "primary"
-            BlueprintIdentifier = "0FFD1B3F1F5E6D2B0007E555"
-            BuildableName = "&#38899;&#12465;&#12441;&#12540;&#65288;&#20206;&#65289;.app"
-            BlueprintName = "&#38899;&#12465;&#12441;&#12540;&#65288;&#20206;&#65289;"
-            ReferencedContainer = "container:&#38899;&#12465;&#12441;&#12540;&#65288;&#20206;&#65289;.xcodeproj">
-         </BuildableReference>
-      </MacroExpansion>
       <AdditionalOptions>
       </AdditionalOptions>
    </TestAction>
@@ -94,16 +65,6 @@
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
-         <BuildableReference
-            BuildableIdentifier = "primary"
-            BlueprintIdentifier = "0FFD1B3F1F5E6D2B0007E555"
-            BuildableName = "&#38899;&#12465;&#12441;&#12540;&#65288;&#20206;&#65289;.app"
-            BlueprintName = "&#38899;&#12465;&#12441;&#12540;&#65288;&#20206;&#65289;"
-            ReferencedContainer = "container:&#38899;&#12465;&#12441;&#12540;&#65288;&#20206;&#65289;.xcodeproj">
-         </BuildableReference>
-      </BuildableProductRunnable>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
